requirements.txt
# Build Frameworks & UI
streamlit==1.52.2
python-dotenv

# LangGraph + LangChain
langgraph
langchain
langchain-core
langchain-community
langchain-text-splitters
langchain-huggingface

# Models & ML Dependencies
huggingface-hub
transformers
torch
torchvision

# Vector DB
faiss-cpu
sentence-transformers

# Voice Features
audio-recorder-streamlit==0.0.10
SpeechRecognition
gTTS
langdetect